requirements.txt
aiohttp
aiodns
beautifulsoup4
cchardet
elasticsearch
pysocks
pandas
aiohttp_socks
schedule
geopy
fake-useragent
twint
seaborn
matplotlib
numpy
scipy
scikit-learn
mxnet
gluonts
jupyterlab
finnhub-python
flask
spacytextblob